Updated: Addressable[..] will now always return UInt64

+++ Sources/SimulatorLib/Units/Addressable.swift
@@ -11,9 +11,9 @@ import Foundation
 public protocol Addressable {
     var data: Data { get set }
 
-    subscript (b b: UInt64) -> UInt8 { get set }
-    subscript (w w: UInt64) -> UInt16 { get set }
-    subscript (l l: UInt64) -> UInt32 { get set }
+    subscript (b b: UInt64) -> UInt64 { get set }
+    subscript (w w: UInt64) -> UInt64 { get set }
+    subscript (l l: UInt64) -> UInt64 { get set }
     subscript (q q: UInt64) -> UInt64 { get set }
 }
 
@@ -24,24 +24,24 @@ extension Addressable {
         }
     }
 
-    public subscript (b b: UInt64) -> UInt8 {
+    public subscript (b b: UInt64) -> UInt64 {
         get {
             let b = Int(b)
             validateAccess(at: b)
-            return data[b]
+            return UInt64(data[b])
         }
         set {
             let b = Int(b)
             validateAccess(at: b)
-            data[b] = newValue
+            data[b] = UInt8(newValue)
         }
     }
-    public subscript (w w: UInt64) -> UInt16 {
+    public subscript (w w: UInt64) -> UInt64 {
         get {
             let w = Int(w)
             validateAccess(at: w + 1)
-            return (UInt16(data[w + 0]) << 0x00) +
-                (UInt16(data[w + 1]) << 0x08)
+            return (UInt64(data[w + 0]) << 0x00) +
+                (UInt64(data[w + 1]) << 0x08)
         }
         set {
             let w = Int(w)
@@ -50,14 +50,14 @@ extension Addressable {
             data[w + 1] = UInt8((newValue & 0xff00) >> 0x08)
         }
     }
-    public subscript (l l: UInt64) -> UInt32 {
+    public subscript (l l: UInt64) -> UInt64 {
         get {
             let l = Int(l)
             validateAccess(at: l + 3)
-            return (UInt32(data[l + 0]) << 0x00) +
-                (UInt32(data[l + 1]) << 0x08) +
-                (UInt32(data[l + 2]) << 0x10) +
-                (UInt32(data[l + 3]) << 0x18)
+            return (UInt64(data[l + 0]) << 0x00) +
+                (UInt64(data[l + 1]) << 0x08) +
+                (UInt64(data[l + 2]) << 0x10) +
+                (UInt64(data[l + 3]) << 0x18)
         }
         set {
             let l = Int(l)
